+++ tools/IoTEdgeDevOps/DevOpsLib/ReleaseManagement.cs
@@ -10,6 +10,7 @@ namespace DevOpsLib
     using Newtonsoft.Json;
     using Newtonsoft.Json.Linq;
 
+
     public class ReleaseManagement
     {
         const string ReleasePathSegmentFormat = "{0}/{1}/_apis/release/releases";
